refactor: hook naming fix
