+++ client/src/components/practice/short/useCurrentTyping.ts
